requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.2
drf-yasg==1.20.0
filelock==3.0.12
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.0
promise==2.3
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyshorteners==1.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Rx==1.6.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
